Ball and Ring disappears after Gameplay Loop, Ring and Reticle are at 0,0,0 now, Levelmanager is fixed.

+++ Assets/Scripts/DetectScoring.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class DetectScoring : MonoBehaviour {
 
-    public int scorePerHit = 1;
+    public long scorePerHit = 1;
 
 	private GvrHUD hud;
